Refactor: Remove redundant string parameter in `toBreezeIconsType` and update usages
The `toBreezeIconsType` function in `BreezeIconEnumConverter.kt` was modified to remove the redundant string parameter. This parameter was redundant because the function was already an extension function of the `String` class.

The function's usages in `EditarValorConta.kt` and `CardBreeze.kt` were updated to reflect this change.

Unnecessary import was removed in `DateConverter.kt`.